Generate_cases/requirements.txt
numpy
pandas
scipy
matplotlib
pathlib
PyYAML